move golden to testutils
