+++ lib/integrations/awsoidc/ec2_ssm_iam_config_test.go
@@ -32,7 +32,7 @@ import (
 	ssmtypes "github.com/aws/aws-sdk-go-v2/service/ssm/types"
 	"github.com/stretchr/testify/require"
 
-	"github.com/gravitational/teleport/lib/utils/golden"
+	"github.com/gravitational/teleport/lib/utils/testutils/golden"
 )
 
 func TestEC2SSMIAMConfigReqDefaults(t *testing.T) {
