+++ lib/integrations/awsoidc/idp_iam_config_test.go
@@ -37,7 +37,7 @@ import (
 	"github.com/gravitational/teleport/lib"
 	awslib "github.com/gravitational/teleport/lib/cloud/aws"
 	"github.com/gravitational/teleport/lib/integrations/awsoidc/tags"
-	"github.com/gravitational/teleport/lib/utils/golden"
+	"github.com/gravitational/teleport/lib/utils/testutils/golden"
 )
 
 func TestIdPIAMConfigReqDefaults(t *testing.T) {
